[AUTOCOMPLETERS] Working on prefix autocompleter
